Move download related functionality to {krate,version}::downloads::*
